+++ birthdaytracker/lib/pages/profile_directory.dart
@@ -1,9 +1,9 @@
 import 'dart:collection';
 import 'package:flutter/material.dart';
-import '../models/birthdayprofile.dart';
+import '../models/birthday_profile.dart';
 import '../models/nav_helper.dart';
-import '../widgets/directoryelements.dart';
-import '../models/profilestore.dart';
+import '../widgets/directory_elements.dart';
+import '../models/profile_store.dart';
 
 class ProfileDirectory extends StatefulWidget {
   const ProfileDirectory({super.key});
