+++ birthdaytracker/lib/widgets/delete_profile_button.dart
@@ -1,4 +1,4 @@
-import 'package:birthdaytracker/models/birthdayprofile.dart';
+import 'package:birthdaytracker/models/birthday_profile.dart';
 import 'package:birthdaytracker/models/hive_helper.dart';
 import 'package:birthdaytracker/models/nav_helper.dart';
 import 'package:flutter/material.dart';
