+++ birthdaytracker/lib/widgets/directory_elements.dart
@@ -1,5 +1,5 @@
-import 'package:birthdaytracker/models/birthdayprofile.dart';
-import 'package:birthdaytracker/pages/viewprofile.dart';
+import 'package:birthdaytracker/models/birthday_profile.dart';
+import 'package:birthdaytracker/pages/view_profile.dart';
 import 'package:flutter/material.dart';
 
 abstract class DirectoryElement extends StatelessWidget {
