+++ birthdaytracker/lib/widgets/profile_picture.dart
@@ -1,4 +1,4 @@
-import 'package:birthdaytracker/models/birthdayprofile.dart';
+import 'package:birthdaytracker/models/birthday_profile.dart';
 import 'package:flutter/material.dart';
 
 class ProfilePicture extends StatelessWidget {
